Creating Custom VPC
The first task expected of a user on AWS console is to create a custom VPC. Why create a VPC? A VPC, virtual private cloud allows you deploy AWS resources in your private environment, similar to you local data center. See a VPC as your datacentre. It is custom to you and you determine how and what to deploy in it